+++ console/src/reducers/index.js
@@ -6,9 +6,10 @@ import templateList from "../containers/VisibleTemplateList/reducers";
 import processForm from "../containers/VisibleProcessForm/reducers";
 import process from "../containers/VisibleProcessPage/reducers";
 import processWizard from "../containers/VisibleProcessWizard/reducers";
+import portal from "../containers/VisiblePortalPage/reducers";
 import * as fromSession from "./session";
 
-export default {history, projectList, project, templateList, log, processForm, process, processWizard};
+export default {history, projectList, project, templateList, log, processForm, process, processWizard, portal};
 
 export const getHistoryState = (state) => state.history;
 export const getProjectListState = (state) => state.projectList;
@@ -18,5 +19,6 @@ export const getTemplateListState = (state) => state.templateList;
 export const getProcessFormState = (state) => state.processForm;
 export const getProcessState = (state) => state.process;
 export const getProcessWizardState = (state) => state.processWizard;
+export const getPortalState = (state) => state.portal;
 
 export const getIsLoggedIn = (state) => fromSession.getIsLoggedIn(state.session);